+++ include/C4030.hpp
@@ -20,10 +20,10 @@ public:
 	void		setLink(std::size_t, nts::IComponent &,
 					std::size_t);
 	void		dump() const;
-	int		getMaxPin() const;
+	size_t		getMaxPin() const;
 	
 private:
-	int						_maxPin;
+	size_t						_maxPin;
 	std::vector<std::unique_ptr<IComponent>>	_pin;
 	std::vector<size_t>				_otherPin;
 };
